Bump felles versjon og bytt til renamed ecb klient
