+++ pom.xml
@@ -42,7 +42,7 @@
         <!-- Nav Familie-->
         <prosessering.version>2.20251103090009_f7d38c7</prosessering.version>
         <felles-kontrakter.version>3.0_20251106145307_0419786</felles-kontrakter.version>
-        <felles.version>3.20251105152539_b03ed80</felles.version>
+        <felles.version>3.20251209142024_1cdfbdb</felles.version>
         <eksterne-kontrakter-bisys.version>2.0_20251207221645_82d9e1d</eksterne-kontrakter-bisys.version>
         <familie.kontrakter.stønadsstatistikk>2.0_20251207221645_82d9e1d</familie.kontrakter.stønadsstatistikk>
         <familie.kontrakter.saksstatistikk>2.0_20251207221645_82d9e1d</familie.kontrakter.saksstatistikk>
