+++ src/test/integrasjonstester/kotlin/no/nav/familie/ks/sak/integrasjon/ecb/ECBIntegrationTest.kt
@@ -3,19 +3,20 @@ package no.nav.familie.ks.sak.integrasjon.ecb
 import io.mockk.every
 import io.mockk.mockk
 import io.mockk.verify
-import no.nav.familie.ks.sak.OppslagSpringRunnerTest
 import no.nav.familie.ks.sak.config.DatabaseCleanupService
 import no.nav.familie.ks.sak.integrasjon.ecb.domene.ECBValutakursCacheRepository
-import no.nav.familie.valutakurs.Frequency
-import no.nav.familie.valutakurs.ValutakursRestClient
-import no.nav.familie.valutakurs.domene.ECBExchangeRate
-import no.nav.familie.valutakurs.domene.ECBExchangeRateDate
-import no.nav.familie.valutakurs.domene.ECBExchangeRateKey
-import no.nav.familie.valutakurs.domene.ECBExchangeRateValue
-import no.nav.familie.valutakurs.domene.ECBExchangeRatesData
-import no.nav.familie.valutakurs.domene.ECBExchangeRatesDataSet
-import no.nav.familie.valutakurs.domene.ECBExchangeRatesForCurrency
-import no.nav.familie.valutakurs.domene.toExchangeRates
+import no.nav.familie.valutakurs.ECBValutakursRestKlient
+import no.nav.familie.valutakurs.domene.ecb.ECBValutakursData
+import no.nav.familie.valutakurs.domene.ecb.toExchangeRates
+import no.nav.familie.valutakurs.domene.ecb.Frequency
+import no.nav.familie.valutakurs.domene.sdmx.SDMXExchangeRate
+import no.nav.familie.valutakurs.domene.sdmx.SDMXExchangeRateDate
+import no.nav.familie.valutakurs.domene.sdmx.SDMXExchangeRateKey
+import no.nav.familie.valutakurs.domene.sdmx.SDMXExchangeRateValue
+import no.nav.familie.valutakurs.domene.sdmx.SDMXExchangeRatesDataSet
+import no.nav.familie.valutakurs.domene.sdmx.SDMXExchangeRatesForCurrency
+import no.nav.familie.ks.sak.OppslagSpringRunnerTest
+
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
@@ -26,7 +27,7 @@ import java.time.LocalDate
 
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
 class ECBIntegrationTest : OppslagSpringRunnerTest() {
-    private val ecbClient = mockk<ValutakursRestClient>()
+    private val ecbClient = mockk<ECBValutakursRestKlient>()
 
     @Autowired
     private lateinit var ecbService: ECBService
@@ -81,19 +82,20 @@ class ECBIntegrationTest : OppslagSpringRunnerTest() {
         frequency: Frequency,
         exchangeRates: List<Pair<String, BigDecimal>>,
         exchangeRateDate: String,
-    ): ECBExchangeRatesData =
-        ECBExchangeRatesData(
-            ECBExchangeRatesDataSet(
+    ): ECBValutakursData =
+        ECBValutakursData(
+            SDMXExchangeRatesDataSet(
                 exchangeRates.map {
-                    ECBExchangeRatesForCurrency(
+                    SDMXExchangeRatesForCurrency(
                         listOf(
-                            ECBExchangeRateKey("CURRENCY", it.first),
-                            ECBExchangeRateKey("FREQ", frequency.toFrequencyParam()),
+                            SDMXExchangeRateKey("CURRENCY", it.first),
+                            SDMXExchangeRateKey("FREQ", frequency.toFrequencyParam()),
                         ),
+                        listOf(),
                         listOf(
-                            ECBExchangeRate(
-                                ECBExchangeRateDate(exchangeRateDate),
-                                ECBExchangeRateValue((it.second)),
+                            SDMXExchangeRate(
+                                SDMXExchangeRateDate(exchangeRateDate),
+                                SDMXExchangeRateValue((it.second)),
                             ),
                         ),
                     )
